Refactor all of this

+++ src/main/java/com/brandonjja/jugg/utils/FileManager.java
@@ -5,10 +5,7 @@ import org.bukkit.Bukkit;
 import org.bukkit.World;
 import org.bukkit.WorldCreator;
 
-import java.io.*;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
+import java.io.File;
 
 public class FileManager {
 
@@ -31,72 +28,17 @@ public class FileManager {
     public static void deleteWorld(File path) {
         if (path.exists()) {
             File[] files = path.listFiles();
-            for (File file : files) {
-                if (file.isDirectory()) {
-                    deleteWorld(file);
-                } else {
-                    file.delete();
+            if (files != null) {
+                for (File file : files) {
+                    if (file.isDirectory()) {
+                        deleteWorld(file);
+                    } else {
+                        file.delete();
+                    }
                 }
             }
         }
         path.delete();
-    }
-
-    /**
-     * Copy a world to a new folder
-     **/
-    public static void copyWorld(File source, File target) {
-        try {
-            List<String> ignore = new ArrayList<String>(Arrays.asList("uid.dat", "session.lock")); // session.dat ??
-            if (!ignore.contains(source.getName())) {
-                if (source.isDirectory()) {
-                    if (!target.exists())
-                        target.mkdirs();
-                    String[] files = source.list();
-                    for (String file : files) {
-                        File srcFile = new File(source, file);
-                        File destFile = new File(target, file);
-                        copyWorld(srcFile, destFile);
-                    }
-                } else {
-                    InputStream inputStream = new FileInputStream(source);
-                    OutputStream outputStream = new FileOutputStream(target);
-                    byte[] buffer = new byte[1024];
-                    int inStreamLength;
-                    while ((inStreamLength = inputStream.read(buffer)) > 0)
-                        outputStream.write(buffer, 0, inStreamLength);
-                    inputStream.close();
-                    outputStream.close();
-                }
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    /**
-     * @param absolutePath The path to format (ex. Bukkit.getWorldContainer().getAbsolutePath() )
-     * @return Formatted path (C:\\Users formatted to C:/Users)
-     */
-    private static String formatAbsolutePath(String absolutePath) {
-        char[] charArr = absolutePath.toCharArray();
-        StringBuilder path = new StringBuilder("");
-        for (char character : charArr) {
-            if (character == '.') {
-                continue;
-            } else if (character == '\\') {
-                character = '/';
-            }
-            path.append(character);
-        }
-        return path.toString();
-    }
-
-    public static File getFileFromName(String name) {
-        String path = formatAbsolutePath(Bukkit.getWorldContainer().getAbsolutePath());
-        String worldName = path + name;
-
-        return new File(worldName);
     }
 
     /**
